+++ app/src/main/java/hd/softeer/luckycardgame/CardSectionAdapter.kt
@@ -83,61 +83,3 @@ class CardFrontViewHolder(private val binding: RvItemCardFrontBinding) :
         Log.d("onBind", card.toString())
     }
 }
-
-// RecyclerView ItemDecorator 클래스
-class PlayerCardItemDecorator(private val cardSize: Int) : RecyclerView.ItemDecoration() {
-    override fun getItemOffsets(
-        outRect: Rect,
-        view: View,
-        parent: RecyclerView,
-        state: RecyclerView.State
-    ) {
-        val position = parent.getChildAdapterPosition(view)
-        val itemCount = parent.adapter?.itemCount ?: 0
-        // 카드가 8장일 때 중간에 겹치도록 설정
-        val moveSize = cardSize * 0.25
-
-        if (position < itemCount - 1) {
-            if (itemCount == 7) {
-                outRect.right -= moveSize.toInt()
-            } else if (itemCount == 8) {
-                outRect.right -= moveSize.toInt() + 4
-            }
-        }
-
-        if (itemCount < 7) {
-            outRect.left += 5
-        }
-    }
-}
-
-class SharedCardItemDecorator(private val spanCount: Int) : RecyclerView.ItemDecoration() {
-    override fun getItemOffsets(
-        outRect: Rect,
-        view: View,
-        parent: RecyclerView,
-        state: RecyclerView.State
-    ) {
-        val space = when (spanCount) {
-            5 -> {
-                40
-            }
-
-            4 -> {
-                56
-            }
-
-            else -> {
-                12
-            }
-        }
-        val position = parent.getChildAdapterPosition(view)
-        val column = position % spanCount
-
-        outRect.left = space - column * space / spanCount
-        outRect.right = (column + 1) * space / spanCount
-        if (position >= spanCount) {
-            outRect.top = space
-        }
-    }
-}
